--- Authentication/Classes/LogoutController.php
@@ -1,9 +1,0 @@
-<?php
-
-class LogoutController{
-	public function Logout(){
-		session_start();
-		session_destroy();
-		header('location:login.php');
-	}
-}
