--- configuration.php
@@ -1,10 +1,0 @@
-<?php
-
-trait ConfigurationDatabase
-{
-    private $HOSTNAME = "127.0.0.1";
-    private $USERNAME = "root";
-    private $PASSWORD = "";
-    private $DBNAME = "ecanteen";
-    private $PORT = 3306;
-}
